+++ src/Faker/Address.cs
@@ -17,6 +17,7 @@ namespace RimuTec.Faker {
          const string yamlFileName = "RimuTec.Faker.locales.en.address.yml";
          locale locale = YamlLoader.Read<locale>(yamlFileName);
          _address = locale.en.faker.address;
+         _cityPrefix = _address.CityPrefix;
          _citySuffix = _address.CitySuffix;
          _streetSuffix = _address.StreetSuffix;
       }
@@ -28,6 +29,14 @@ namespace RimuTec.Faker {
       public static string BuildingNumber() {
          var template = _address.BuildingNumber.Random();
          return template.Bothify();
+      }
+
+      /// <summary>
+      /// Resturns a city prefix. Example: "Lake"
+      /// </summary>
+      /// <returns></returns>
+      public static string CityPrefix() {
+         return _address.CityPrefix.Random();
       }
 
       /// <summary>
@@ -137,6 +146,7 @@ namespace RimuTec.Faker {
       }
 
       internal static string[] _streetSuffix;
+      internal static string[] _cityPrefix;
       internal static string[] _citySuffix;
 
       private static address _address;